+++ modules/rosie/rcli/cmd_compile.v
@@ -1,6 +1,8 @@
 module rcli
 
+import os
 import cli
+import rosie
 import rosie.compiler.v2 as compiler
 import rosie.parser.core_0 as parser
 
@@ -8,17 +10,22 @@ pub fn cmd_compile(cmd cli.Command) ? {
 	rosie := init_rosie_with_cmd(cmd) ?
 	debug := 0
 
-	mut pat_str := rosie.rpl + cmd.args[0]
-	fname := cmd.args[1]
+	rplx_fname := cmd.args[1]
 	entrypoints := if cmd.args.len > 2 {
 		cmd.args[2..]
 	} else {
 		["*"]
 	}
 
-	eprintln("rpl: '$pat_str'")
 	mut p := parser.new_parser(debug: debug)?
-	p.parse(data: pat_str)?
+	if os.is_file(cmd.args[0]) == false {
+		pat_str := rosie.rpl + cmd.args[0]
+		//eprintln("rpl: '$pat_str'")
+		p.parse(data: pat_str)?
+	} else {
+		fname := cmd.args[0]
+		p.parse(file: fname)?
+	}
 
 	for e in entrypoints {
 		p.expand(e)?
@@ -29,5 +36,5 @@ pub fn cmd_compile(cmd cli.Command) ? {
 		c.compile(e)?
 	}
 
-	c.rplx.save(fname, true)?
+	c.rplx.save(rplx_fname, true)?
 }